fix bugs; schema registry getting compatibility & refreshing features; refactor some ser-deserializer code

+++ src/main/java/com/example/mytool/ui/control/SchemaEditableTableControl.java
@@ -1,7 +1,9 @@
 package com.example.mytool.ui.control;
 
 import com.example.mytool.manager.SchemaRegistryManager;
+import com.example.mytool.model.kafka.SchemaMetadataFromRegistry;
 import com.example.mytool.ui.SchemaTableItem;
+import com.example.mytool.ui.util.ViewUtil;
 import io.confluent.kafka.schemaregistry.client.SchemaMetadata;
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
 import javafx.beans.property.SimpleStringProperty;
@@ -18,19 +20,22 @@ import java.util.List;
 
 @Slf4j
 public class SchemaEditableTableControl extends EditableTableControl<SchemaTableItem> {
+    private String selectedClusterName;
 
-    public void setItems(List<SchemaMetadata> schemaMetadataList, String clusterName) {
+    public void setItems(List<SchemaMetadataFromRegistry> schemaMetadataList, String clusterName) {
         ObservableList<SchemaTableItem> items = FXCollections.observableArrayList(schemaMetadataList.stream().map(schemaMetadata -> mapFromSchemaMetaData(schemaMetadata, clusterName)).toList());
         tableItems.setAll(items);
+        selectedClusterName = clusterName;
     }
 
-    static SchemaTableItem mapFromSchemaMetaData(SchemaMetadata schemaMetadata, String clusterName) {
+    static SchemaTableItem mapFromSchemaMetaData(SchemaMetadataFromRegistry schemaMetadataFromRegistry, String clusterName) {
+        SchemaMetadata schemaMetadata = schemaMetadataFromRegistry.schemaMetadata();
         return new SchemaTableItem(
                 schemaMetadata.getSubject(),
                 schemaMetadata.getId(),
                 schemaMetadata.getVersion(),
                 schemaMetadata.getSchemaType(),
-                "UNKNOW_YET",
+                schemaMetadataFromRegistry.compatibility(),
                 schemaMetadata.getSchema(),
                 clusterName
         );
@@ -51,16 +56,26 @@ public class SchemaEditableTableControl extends EditableTableControl<SchemaTable
     }
 
     @FXML
+    public void refresh() throws RestClientException, IOException {
+        if (this.selectedClusterName != null) {
+            List<SchemaMetadataFromRegistry> schemaMetadataList = SchemaRegistryManager.getInstance().getAllSubjectMetadata(this.selectedClusterName);
+            setItems(schemaMetadataList, this.selectedClusterName);
+        }
+    }
+
+    @FXML
     protected void removeItem() {
         List<Integer> indicesToRemove = table.getSelectionModel().getSelectedIndices().reversed();
         indicesToRemove.forEach((i) -> {
             SchemaTableItem item = tableItems.get(i);
-            try {
-                SchemaRegistryManager.getInstance().deleteSubject(item.getClusterName(), item.getSubject());
-            } catch (RestClientException | IOException e) {
-                throw new RuntimeException(e);
+            if (ViewUtil.confirmAlert("Delete Subject", "Are you sure to delete " + item.getSubject() + " ?", "Yes", "Cancel")) {
+                try {
+                    SchemaRegistryManager.getInstance().deleteSubject(item.getClusterName(), item.getSubject());
+                } catch (RestClientException | IOException e) {
+                    throw new RuntimeException(e);
+                }
+                tableItems.remove((int) i);
             }
-            tableItems.remove((int) i);
         });
     }
